@@ -12,7 +12,7 @@
         <maven.compiler.target>17</maven.compiler.target>
         <solace-messaging-client.version>1.4.0</solace-messaging-client.version>
         <solclientj.version>10.0.0</solclientj.version>
-        <spring-boot.version>3.3.10</spring-boot.version>
+        <spring-boot.version>3.4.4</spring-boot.version>
         <jupiter.version>5.10.2</jupiter.version>
         <camel.version>4.8.3</camel.version>
         <commons-collections4.version>4.4</commons-collections4.version>
