@@ -10,9 +10,9 @@ import com.solace.maas.ep.event.management.agent.plugin.command.model.ExecutionT
 import com.solace.maas.ep.event.management.agent.plugin.mop.MOPSvcType;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.boot.test.mock.mockito.SpyBean;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.context.bean.override.mockito.MockitoSpyBean;
 
 import java.util.List;
 
@@ -26,12 +26,12 @@ import static org.mockito.Mockito.verifyNoInteractions;
 @ActiveProfiles("TEST")
 class CommandMessageProcessorTests {
 
-    @MockBean
+    @MockitoBean
     private CommandManager commandManager;
-    @SpyBean
+    @MockitoSpyBean
     private CommandMessageProcessor commandMessageProcessor;
 
-    @MockBean
+    @MockitoBean
     private DynamicResourceConfigurationHelper dynamicResourceConfigurationHelper;
 
     @Test
